+++ lib/Controller/Configuration/CurrencyController.php
@@ -9,11 +9,14 @@ class CurrencyController extends AbstractCrudController
     protected function customData( Request $request, $entity = null ): array
     {
         $currencies = [];
-        foreach ( $this->resources as $currency ) {
-            $currencies[$currency->getCode()]   = [
-                'symbol'    => Currencies::getSymbol( $currency->getCode() ),
-                'name'      => Currencies::getName( $currency->getCode() ),
-            ]; 
+        
+        if ( $this->resources ) {
+            foreach ( $this->resources as $currency ) {
+                $currencies[$currency->getCode()]   = [
+                    'symbol'    => Currencies::getSymbol( $currency->getCode() ),
+                    'name'      => Currencies::getName( $currency->getCode() ),
+                ]; 
+            }
         }
         
         return [